requirements.txt
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-codicefiscale==0.10.3
python-dateutil==2.9.0.post0
python-fsutil==0.15.0
python-slugify==8.0.4
six==1.17.0
text-unidecode==1.3
